+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Ampache: A few more actions now support pagination with offset and limit: `artist_albums`, `artist_songs`, `album_songs`, `search_songs`
 - Subsonic: Added support to `getArtistInfo` to identify the artist using a track, an album, or a folder ID
   [#906](https://github.com/owncloud/music/issues/906)
+- Subsonic: Added support to `getAlbumInfo` to identify the album using a track or folder ID
 
 ### Fixed
 - A performance problem affecting Subsonic method `getArtist`, Ampache action `artist_albums`, and a few other functions